Issue #1375: Created modules .building.osx and building.templates.
