+++ PyInstaller/building/makespec.py
@@ -12,126 +12,19 @@
 Automatically build spec files containing a description of the project
 """
 
-
 import os
 import sys
 
-
 from PyInstaller import HOMEPATH, DEFAULT_SPECPATH
 from PyInstaller import log as logging
+from PyInstaller.building.templates import onefiletmplt, onedirtmplt, comsrvrtmplt, cipher_absent_template, \
+    cipher_init_template, bundleexetmplt, bundletmplt
 from PyInstaller.compat import expand_path, is_win, is_cygwin, is_darwin
 
 logger = logging.getLogger(__name__)
 
 # HACK to be used used in tests
 _EXENAME_FORCED_SUFFIX_ = ''
-
-onefiletmplt = """# -*- mode: python -*-
-%(cipher_init)s
-
-a = Analysis(%(scripts)s,
-             pathex=%(pathex)s,
-             hiddenimports=%(hiddenimports)r,
-             hookspath=%(hookspath)r,
-             runtime_hooks=%(runtime_hooks)r,
-             excludes=%(excludes)s,
-             cipher=block_cipher)
-pyz = PYZ(a.pure,
-             cipher=block_cipher)
-exe = EXE(pyz,
-          a.scripts,
-          a.binaries,
-          a.zipfiles,
-          a.datas,
-          name='%(exename)s',
-          debug=%(debug)s,
-          strip=%(strip)s,
-          upx=%(upx)s,
-          console=%(console)s %(exe_options)s)
-"""
-
-onedirtmplt = """# -*- mode: python -*-
-%(cipher_init)s
-
-a = Analysis(%(scripts)s,
-             pathex=%(pathex)s,
-             hiddenimports=%(hiddenimports)r,
-             hookspath=%(hookspath)r,
-             runtime_hooks=%(runtime_hooks)r,
-             excludes=%(excludes)s,
-             cipher=block_cipher)
-pyz = PYZ(a.pure,
-             cipher=block_cipher)
-exe = EXE(pyz,
-          a.scripts,
-          exclude_binaries=True,
-          name='%(exename)s',
-          debug=%(debug)s,
-          strip=%(strip)s,
-          upx=%(upx)s,
-          console=%(console)s %(exe_options)s)
-coll = COLLECT(exe,
-               a.binaries,
-               a.zipfiles,
-               a.datas,
-               strip=%(strip)s,
-               upx=%(upx)s,
-               name='%(name)s')
-"""
-
-comsrvrtmplt = """# -*- mode: python -*-
-%(cipher_init)s
-
-a = Analysis(%(scripts)s,
-             pathex=%(pathex)s,
-             hiddenimports=%(hiddenimports)r,
-             hookspath=%(hookspath)r,
-             runtime_hooks=%(runtime_hooks)r,
-             excludes=%(excludes)s,
-             cipher=block_cipher)
-pyz = PYZ(a.pure,
-             cipher=block_cipher)
-exe = EXE(pyz,
-          a.scripts,
-          exclude_binaries=True,
-          name='%(exename)s',
-          debug=%(debug)s,
-          strip=%(strip)s,
-          upx=%(upx)s,
-          console=%(console)s %(exe_options)s)
-dll = DLL(pyz,
-          a.scripts,
-          exclude_binaries=True,
-          name='%(dllname)s',
-          debug=%(debug)s)
-coll = COLLECT(exe, dll,
-               a.binaries,
-               a.zipfiles,
-               a.datas,
-               strip=%(strip)s,
-               upx=%(upx)s,
-               name='%(name)s')
-"""
-
-cipher_absent_template = """
-block_cipher = None
-"""
-
-cipher_init_template = """
-block_cipher = pyi_crypto.PyiBlockCipher(key=%(key)r)
-"""
-
-bundleexetmplt = """app = BUNDLE(exe,
-             name='%(exename)s.app',
-             icon=%(icon)s,
-             bundle_identifier=%(bundle_identifier)s)
-"""
-
-bundletmplt = """app = BUNDLE(coll,
-             name='%(name)s.app',
-             icon=%(icon)s,
-             bundle_identifier=%(bundle_identifier)s)
-"""
 
 
 def quote_win_filepath(path):
